radare2: Switch back to Intel syntax
Phew, for some (weird) reason I'm used to using AT&T syntax in objdump
and can't for the life of me get used to Intel syntax in radare2.

Especially when being deeply zoned into a reverse-engineering session,
AT&T syntax is annoying me so much that I immediately switch back to
Intel syntax.

Maybe I should do that with objdump, GDB and others as well, but since I
usually do longer sessions (which may be the reason why it's not so much
bugging me in objdump) using radare2 let's at least get that nuisance
out of the way.

Signed-off-by: aszlig <aszlig@nix.build>