modules/sternenseemann/desktop-sway: use color scheme in mako
Also move notifications to the bottom right corner with appropriate
padding from the bottom of the screen.

Move my Plan9-esque color scheme to the desktop-sway module to make
reusing the colors a little more intelligible. The default colors of the
sway module are somewhat close to i3's default colors now instead.