vim: Add support for EditorConfig
I use this in most of my projects at work where people not necessarily
use Vim for editing files, so that there is a common standard eg. for
indentation.

Since I'm also the author of these .editorconfig files, it usually is
not an issue, but given that I expect other people to use
plugins/editors that recognise the format, I myself should do that as
well (and hereby it's done).

Signed-off-by: aszlig <aszlig@nix.build>