Partially revert "machines/sternenseemann/wolfgang: install ad"
This partially reverts commit a61f622798071e0b779da28f471299c5b0387891.

Reason for revert: The released version of ad is not really ready to be
used by me since it is broken in a variety of ways on wayland (way to go
for a TUI application…). I'll just use ad from a local checkout somehow
in the mean time since I've also had a hard time getting fixes
upstreamed.