pkgs/sternenseemann/scripts: add nix-instantiate-to
Adds a shortcut script which instantiates derivations and immediately
copies them to a remote host. I use this extensively for testing builds
on a faster remote machine instead of the Nix remote builder protocol.
The disadvantage of the latter is that every intermediate build product
needs to be copied to the local store which I'm seldomly interested in.
In most cases, I just want to know whether the build succeeds.