pkgs/sternenseemann/scripts/nix-build-on: support --check
Since --check can't be passed to nix-instantiate(1), we need to
implement arg splitting which is a little tricky. For
nix-instantiate(1), just using a bash array (to prevent arguments from
being split) is enough. Since we call nix-store(1) via ssh, the shell on
the remote host will interpret the string again, so we need to escape
the arguments beforehand (though in practice, none of the flags
currently have spaces in them).