machines/sternenseemann/wolfgang: add zathurarc
The use of the primary selection of zathura is not super useful and a
bit annoying. The double g bind to jump to the top is a little bit
unnecessary since there is no other bind with the g- prefix (I often
have zathura stuck in g-… forever for now reason when I hit g
accidentally).