git: Switch to using histogram diff algorithm
Some of the things I frequently do is using "git diff --histogram" and
given that most (or always?) of the time the diffs are more readable I
think I should give it a try to set it by default.

If it turns out that I'm longing for "git diff --diff-algorithm=myers"
at some point, I can still turn it back to the default but I'm pretty
sure that it's unlikely to happen.

Signed-off-by: aszlig <aszlig@nix.build>