git: Default to verbose commit editing style
When I create a commit, I *always* use "git commit -vs" (the -s is for
sign-off) and it's so deep in my muscle memory that I don't think about
it anymore.

That said, I also had -S (for --gpg-sign) in my muscle memory, so
getting rid of one more letter is something that's probably worth it,
even though it might take a while to adapt.

Signed-off-by: aszlig <aszlig@nix.build>