vim: Fix highlighting line len with EditorConfig
By default, the editorconfig plugin highlights the max line length by
highlighting the boundary column on every line.

However, we already highlight this as part of the normal configuration,
the only difference being that I originally hardcoded a 79 textwidth.

Using the configured textwidth in conjunction with disabling the
highlighting via the editorconfig plugin now not only deduplicates
functionality but I also get highlighting as I'm used to, highlighting
only the columns that exceed the textwdith.

Signed-off-by: aszlig <aszlig@nix.build>