tests/psi: Switch to evilwm instead of icewm
I haven't found ot what's the exact change in IceWM, but version 3.7.4
lead to a few visual changes that make it particularly hard to OCR,
which lead to the test timing out.

Since EvilWM is a lot more minimal (especially when it comes to
colours), the OCR result is not only much better but also quicker to
obtain.

Signed-off-by: aszlig <aszlig@nix.build>