vim: Don't use EditorConfig for diffs
While in theory there could be repositories that want an EditorConfig to
be applied to a diff file, I couldn't come up with a *practical* reason
to do so.

However, when using git with interactive hunk selection mode and it's
edit feature, enabling an EditorConfig that applies to "*" is very
counterproductive and could result in invalid diffs.

Since I'm at it, I also excluded diffs from my "line is too long"
highlighting, since it doesn't usually make sense there either.

Signed-off-by: aszlig <aszlig@nix.build>