pkgs/sternenseemann/nix-instantiate-on: use nix-store --export
I've known that this is faster for a long time, but never changed the
script out of laziness. For typical cases (100-200 drv files), I've seen
something in the realm of a 5 times speedup. Avoiding the overhead of
the individual ssh connections per path does indeed seem to worth it.
It's still not exactly fast, but much more tolerable.

I've also tested compression a bit. If your connection is not the
bottleneck, it doesn't seem to help except in pathological
cases (copying big source directories), but the cost is laughable
compared to the nix-store(1) commands.