core/tests: Properly remove ZSH test
Commit 80f297358f0bd0c62f3b0523a09deea1fc0e5e2e tried to remove that
test because evaluation has probably failed at some point. To remove the
test however we need to remove the *whole* check attribute set instead
of just a component of the path.

Removing just the final (attribute) path component however leads to
evaluation of the full top-level attribute of nixpkgs and we get an eval
error like this (indented for better readability):

  error: cannot coerce a set to a string: {
    _3proxy = «thunk»;
    aaaaxy = «thunk»;
    acme = «thunk»;
    acme-dns = «thunk»;
    activation = «thunk»;
    activation-etc-overlay-immutable = «thunk»;
    activation-etc-overlay-mutable = «thunk»;
    activation-lib = «thunk»;
    activation-nix-channel = «thunk»;
    activation-perlless = «thunk»;
    «1136 attributes elided»
  }

Signed-off-by: aszlig <aszlig@nix.build>