pkgs/sternenseemann/nix-instantiate-to: don't chunk args to --export
xargs(1) can split up the arguments to multpile commands to migitate
e.g. ARG_MAX. However, we don't want that since the output of
nix-store(1) --export is not concatenatable and we don't know that any
subsets of its output are importable as is (since there may be dangling
references then). Consequently, we can just use splitting in normal bash
knowing that it limits our maximum closure size.

To properly fix this issue, we'd need either concatenatable exports or
nix-store(1) --export that calculates closures for itself or reads paths
from stdin.