sandbox: Port nix-query.cc to newer Nix versions
A lot has changed around the Nix store abstraction between Nix 2.3 and
current Nix versions, so that keeping track of the versions in between
is too much for a time-constrained "let's just make it work with current
Nix".

This means that I've trown out backward-compatibility with Nix 2.3 and
older and just moved on with querying the store in the current version.

Since currently my machines do not evaluate due to Nix 2.3 being marked
as insecure, I was about to "quick fix" this by simply overriding the
evaluation error by allowing use of Nix 2.3.

However, I do use the sandbox implementation in other projects and I
just happened to stumble on another issue:

When running Nix as root, the store initialisation from Nix tries to fix
permissions of the /nix/store directory, which fails if the store is
bind-mounted read-only.

Debugging and/or fixing that issue with an ancient Nix version is not a
lot of fun, so that's why I finally decided to port it properly.

The above issue might even no longer be the case with newer Nix
versions, so that would be another reason.

Signed-off-by: aszlig <aszlig@nix.build>