pkgs/sternenseemann/tep: replace tepData.hs with small awk script
tepData.hs uses a general purpose parser for emoji-test.txt I've
implemented for the emoji-generic library. This parser tries to
completely preserve all information in the data file and has a rich data
type to represent emojis, but not much more than that. The project has
been on ice for years now. I still think the ideas I had for it are
worth pursuing, but other projects are more pressing and I don't have
any concrete use case for emoji-generic at the moment.

Its parser is overkill for tep. For generating the emojis.txt file for
tep, we only need to keep track of group and subgroup, pull out the CLDR
short name and the actual emoji of every line and format the result.
This can easily be implemented using gawk(1) as shown here.

Migrating tep to the AWK script frees me from porting emoji-generic to
new versions of GHC as Nixpkgs upgrades and allows me to archive the
repository for now. Hopefully a reason to resurrect it will eventually
present itself.

pkgs/sternenseemann/emoji-generic: remove