modules/sternenseemann/ma: use PATH from environment block
systemd services in NixOS implicitly set a default path which is a huge
nuisance and very confusing since it overwrites whatever is set in the
systemd environment block. The only way to disable is to force the path
to the empty list (which does not result in an empty PATH!).

ma should have the normal user environment. To achieve this with a
systemd user service, the normal environment needs to be imported with
systemctl(1) --user import-environment. Luckily for me, niri-session
does this automatically.