dnyarri: Inline gPodder and use nixpkgs version
I initially added this in Vuizvui because I wanted to have support for
yt-dlp instead of youtube-dl. This is now also the default in nixpkgs,
so there is no reason to keep this around.

The reason why I'm doing this now is because we've run into another eval
error due to Python version mismatches:

  error: Python version mismatch in 'gpodder-git-2023-07-24':

  The Python derivation 'gpodder-git-2023-07-24' depends on a Python
  derivation named 'python3.11-dbus-python-1.4.0', but the two
  derivations use different versions of Python:

        'gpodder-git-2023-07-24' uses /nix/store/iyff8129iampdw13nlfqalzhxy8y1hi9-python3-3.13.6
  'python3.11-dbus-python-1.4.0' uses /nix/store/9bzwn5qvvgmczx0ilxacisf7cqiqlz9j-python3-3.11.13

Signed-off-by: aszlig <aszlig@nix.build>