modules/sternenseemann/desktop: disable animations
This makes using niri feel much more snappy if your workflow includes
opening a lot of window which is the case when using the mouse to
navigate in ma(1).

Niri relies on animations to keep you oriented in the virtual
screenspace, but I've found that the brain is able to fill in the gaps
quite well unless you have a lot of windows that look the same next to
each other which makes it look a little strange. However, since you are
still initiating the actions explicitly, you stay oriented in general.
What is more confusing, though, is jumping around with Mod+Tab.

What is terrible without animations is touchpad navigation since it will
gradually move the screen to match your movement on the touchpad, but
snap into place as soon as you lift the fingers off the touchpad. So
essentially, you manually animate the motion for ⅓ of it and the last ⅔
of the motion is performed instantly which is jarring and uncomfortable.
It seems to me that niri would need to add a feature to gate animations
on the type of action that initiated them. See
https://github.com/YaLTeR/niri/discussions/2360.