aszlig/vim: Fix regression in rustfmt behaviour
I've bisected this to the Vim update from version 9.1.1566 to 9.1.1623
and in particular the following Vim commit:

https://github.com/vim/vim/commit/1afe8c3a4d9d61a4762c11ddf4c50050a92bd23e

The commit in question tries to improve loading times by avoiding to
search for rustfmt.toml and by avoiding to check which version of
rustfmt is in scope.

While the latter is something that could be avoided in my setup since
rustfmt is provided by the Nix store and always stays the same, the
former however is something I *do* want, since I work in different Rust
projects with different rustfmt settings.

Signed-off-by: aszlig <aszlig@nix.build>