machines/sternenseemann: use ma as editor and “terminal emulator”
fish uses ANSI escape sequences heavily, so I had to switch to bash
which is alright even in a more feature complete video terminal emulator
like foot (which I have as a fallback) thanks to the work done in
3d7b4f56.

win currently beats out all alternatives because it is able to integrate
with ma via plumb, but this could presumably be added to other software
like saneterm.

This is not exactly a setup I would recommend to anyone, but I'm slowly
working towards something I really like, both by working on my fork of
ma and figuring out how to configure (the) software around it.