tests/psi: Revert switching to evilwm
This reverts commit b29afd1dbea0d1cbd03337ffaf69d874f0dc0baf.

I've bisected the regression to the following nixpkgs commit:

https://github.com/NixOS/nixpkgs/commit/f8a26c443a08644b37d0775bd637bc0901959190

The commit in question has changed the path to
$out/share/fonts/X11/misc/fonts.alias and so EvilWM bails with the
following message when starting:

  couldn't find a font to use: try starting with -fn fontname

I have no idea whether this is an issue with EvilWM, with the nixpkgs
change or my configuration. Since the test isn't about EvilWM or
anything Xorg-related and I don't have time to properly debug this, I
decided to just switch to the default common/x11.nix for the test for
now, since switching to EvilWM was a workaround for OCR issues in the
first place.

Signed-off-by: aszlig <aszlig@nix.build>